Crate ecclesExperimental

An Entity Component Processor implemented in Rust

Based on concepts detailed at this wiki

Macros

aspect!
Make an aspect from the given filter

bag!
Efficiently construct a bag with the given values in

comp!
Define a component with the given fields and default values

entity!
Create a new entity with the given components in

world!
Make a world from the given components and processors

Structs

Bag
Collection type a bit like Vec that avoids generics yet is still type-safe

It has a vector of bytes and allocates values to it on-the-fly. It stores the vtable of the type so it can drop values when they are removed.

Examples

ComponentManager
Attaches component types to the component lists

ComponentType
Identifies the type of component

EntityManager
Manages entity allocation and deletion

EntityRef
An immutable reference into a world with an entity This is pretty much equivalent to a MetaEntity in most ECS implementations

EntityRefMut
An immutable reference into a world with an entity

IntervalProcessor
A processor that runs every interval instead of every tick.

ProcessorManager
Encapsulates processors and their categories

World
Encapsulates the entities, the components they consist of and the processors that operate upon them

Traits

Aspect
An Aspect is used by processors as a matcher against entities, to check if a processor is interested in an entity

Component
All the data for one aspect of the object

EntityBuilder
A closure capably of building an entity

IntoBag
Provides efficient conversion into a bag

Processor
Processes components of entities With each Processor running continuously as if it has a private internal thread, performing actions on every Entity that possesses a Component of a matching aspect to the Processor

Type Definitions

Entity
A unique ID that tags each game object as a separate item Implementations typically use a plain integer for this, as is used here