Struct eccles::BagExperimental

pub struct Bag {
    // some fields omitted
}

Collection type a bit like Vec that avoids generics yet is still type-safe

It has a vector of bytes and allocates values to it on-the-fly. It stores the vtable of the type so it can drop values when they are removed.

Examples

let mut bag = Bag::new();
bag.add(1i);
bag.add(2i);
assert_eq!(vec.len(), 2);
assert_eq!(vec[0], 1);

There is a bag! macro to streamline bag initilisation. You can use it like so: let bag = bag![1u, 2u, 3u, 5u, 7u, 11u]; assert_eq(bag.len(), 6);

Methods

impl Bag

fn new<T: 'static>() -> Bag

Make a new bag of items

fn with_capacity<T: 'static>(capacity: uint) -> Bag

Make a bag of items allocated to allow capacity of T before requiring re-allocation.

fn from_elem<T: 'static + Clone>(length: uint, value: T) -> Bag

Construct a bag by filling it with length copies of value

fn get<T: Copy + 'static>(&self, index: uint) -> T

Get the copy of the value at a given index

fn capacity(&self) -> uint

Get the maximum number of items this bag can hold before reallocating

fn len(&self) -> uint

Get the number of items in this bag

fn is_empty(&self) -> bool

Check if this bag is empty

fn clear(&mut self)

Clear this bag of all the items in it

fn as_slice<T: 'static>(&self) -> &[T]

Return this bag as a slice

fn add<T: 'static>(&mut self, component: T) -> uint

Add a component to the bag

fn into_vec<T: 'static>(self) -> Vec<T>

Convert the bag into a vector of items