Struct eccles::ComponentManagerExperimental

pub struct ComponentManager {
    // some fields omitted
}

Attaches component types to the component lists

Components are stored as described in this article But they're stored in the ComponentManager here because that makes much more sense

Methods

impl ComponentManager

fn new() -> ComponentManager

Create a new component manager

fn with_capacity(capacity: uint) -> ComponentManager

Create a new component manager

fn has_id(&self, entity: Entity, id: &ComponentType) -> bool

Check if the components has a component attached to the entity

fn has<T: Component>(&self, entity: Entity) -> bool

Check if the components has a component attached to the entity

fn add<T: Component>(&mut self, entity: Entity, component: T)

Add a component attached to the entity

fn set<T: Component>(&mut self, entity: Entity, component: T)

Set the component in the components list, and attach it to the entity

fn borrow<T: Component>(&self, entity: Entity) -> Option<&T>

Borrow immutably the component from the components list attached to this entity

fn borrow_mut<T: Component>(&mut self, entity: Entity) -> Option<&mut T>

Borrow mutably the component from the components list attached to this entity

fn get_list<T: Component>(&mut self) -> &mut ComponentList

Get the components list for a component

fn get<T: Component + Clone>(&self, entity: Entity) -> Option<T>

Clone the component from the components list attached to this entity

fn remove<T: Component>(&mut self, entity: Entity)

Remove the component in the components list, and attach it to the entity

fn clear_entity(&mut self, entity: Entity)

Remove all the components in the component lists attached to the entity

fn clear<T: Component>(&mut self)

Clear all the components in a component list

fn register<T: Component>(&mut self)

Register a component