Struct eccles::EntityRefExperimental

pub struct EntityRef<'a> {
    pub components: &'a ComponentManager,
    pub entities: &'a EntityManager,
    pub entity: Entity,
}

An immutable reference into a world with an entity This is pretty much equivalent to a MetaEntity in most ECS implementations

Fields

components

The components of the world the entity resides in

entities

The entities of the world the entity resides in

entity

The entity, a UUID

Methods

impl<'a> EntityRef<'a>

fn is_active(&self) -> bool

Check if the entity is active

fn get<T: Component + Clone>(&mut self) -> T

Clone the component of the given type attached to this entity

fn has<T: Component>(&mut self) -> bool

Check if any components of the given type are attached to this entity

fn has_id(&mut self, id: &ComponentType) -> bool

Check if any components of the given type are attached to this entity

fn borrow<T: Component>(&mut self) -> Option<&'a T>

Borrow the component of the given type attached to this entity