Struct eccles::IntervalProcessorExperimental

pub struct IntervalProcessor<S> {
    pub interval: f64,
    pub since_last: f64,
    pub processor: S,
}

A processor that runs every interval instead of every tick.

Fields

interval

The interval, in seconds

since_last

The number of seconds since the last update

processor

The actual processor to run

Methods

impl<S> IntervalProcessor<S> where S: Processor

fn new(processor: S, interval: Duration) -> IntervalProcessor<S>

Makes a new interval processor from the given processor and interval