Struct eccles::WorldExperimental

pub struct World {
    pub components: ComponentManager,
    pub processors: ProcessorManager,
    pub entities: EntityManager,
}

Encapsulates the entities, the components they consist of and the processors that operate upon them

Fields

components

Manages the data

processors

Manages the processors of the data

entities

Manages the unique identifiers of bits of data

Methods

impl World

fn new() -> World

Creates a new world

fn with_capacity(capacity: uint) -> World

Creates a new world

fn new_entity(&mut self) -> EntityRefMut

Creates an entity in the world and returns it and the world

fn build_entity<B: EntityBuilder>(&mut self, builder: B) -> Entity

Build an entity and enable it

fn build_entities<B: EntityBuilder>(&mut self, count: uint, builder: B) -> Vec<Entity>

Build a certain number of entities and enable them

fn remove_entity(&mut self, entity: Entity)

Removes an entity from the world

fn get_entity(&self, entity: Entity) -> EntityRef

Get an immutable reference to an entity in the world

fn get_entity_mut(&mut self, entity: Entity) -> EntityRefMut

Get a mutable reference to an entity in the world

fn update(&mut self, delta: f64)

Update all the processors with a delta, in seconds

fn begin_loop(self)

Begin a loop of updating the world

fn register_component<C: Component>(&mut self)

Allocate a component list for the given type of component

fn register_processor(&mut self, processor: Box<Processor>, aspect: Box<Aspect>)

Add a processor to the world

fn entities(&self) -> EntityIter

Iterate through the entities