Trait eccles::AspectExperimental

pub trait Aspect: 'static {
    fn check<'a>(&mut self, entity: EntityRef<'a>) -> bool;
}

An Aspect is used by processors as a matcher against entities, to check if a processor is interested in an entity

Required Methods

fn check<'a>(&mut self, entity: EntityRef<'a>) -> bool

Check if the entity meets the criteria of the aspect