Trait eccles::ProcessorExperimental

pub trait Processor: 'static {
    fn run(&mut self, entity: EntityRefMut, delta: f64);

    fn run_all(&mut self, entities: &mut EntityManager, components: &mut ComponentManager, entity: &Vec<Entity>, delta: f64) { ... }
}

Processes components of entities With each Processor running continuously as if it has a private internal thread, performing actions on every Entity that possesses a Component of a matching aspect to the Processor

Required Methods

fn run(&mut self, entity: EntityRefMut, delta: f64)

Process a single entity with a delta in seconds

Provided Methods

fn run_all(&mut self, entities: &mut EntityManager, components: &mut ComponentManager, entity: &Vec<Entity>, delta: f64)

Process entities with a delta in seconds